refactor: 💡 remove redundancy in api import name
